separating into parallel v non parallel
