+++ non-parallel/triangle_corr.txt
@@ -1,0 +1,10 @@
+4.106125235237188809e-02
+2.501939274274746494e-01
+5.010376637981939218e-02
+3.342020361000729717e-02
+-2.822595965330366649e-03
+-1.774910133658860441e-03
+2.873442551639104198e-02
+7.412099432451421666e-02
+8.205197311334211374e-02
+8.751240528360282322e-02
