+++ parallel/bubbles.txt.txt
@@ -1,0 +1,10 @@
+4.074338642047124759e-02
+5.426124531400268181e-01
+8.678075026038550854e-02
+7.369574917774546208e-02
+5.625967390500114657e-02
+2.889775946467814868e-02
+4.307972131537373545e-02
+4.012359904278554878e-02
+5.767399935844246678e-02
+7.204831890265889216e-02
