+++ parallel/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+echo 'bubbles.txt' | mpiexec -n 8 python triangleCorrelation.py
