--- run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-echo 'bubbles.txt' | mpiexec -n 2 python triangleCorrelation.py
